+++ assets/src/scripts/components/graph/state/index.js
@@ -1,8 +1,10 @@
 import cursor from './cursor';
 import layout from './layout';
+import options from './options';
 
 export { getCursor, getCursorDatum, setCursor } from './cursor';
 export * from './layout';
+export * from './options';
 export { getActiveClasses, getChartPoints, getCurrentWaterLevels,
          getCurrentWaterLevelUnit, getExtentX, getLineSegments } from './points';
 export * from './scales';
@@ -10,5 +12,6 @@ export * from './well-log';
 
 export default {
     ...cursor,
-    ...layout
+    ...layout,
+    ...options
 };