+++ assets/src/scripts/components/graph/view/lithology.js
@@ -20,8 +20,7 @@ export const rgbFromHex = function (hexString) {
     ];
 };
 
-export default function (elem, {lithology, selectedLithologyId}, store, siteKey, container) {
-    console.log(container);
+export default function (elem, {lithology, visible, selectedLithologyId}, store, siteKey, container) {
     container = container || elem
         .append('g')
             .classed('lithology', true);
@@ -29,6 +28,12 @@ export default function (elem, {lithology, selectedLithologyId}, store, siteKey,
     // Remove any previously drawn children
     container.selectAll('*').remove();
 
+    // If we're not visible, return before adding lithology layers
+    if (!visible) {
+        return container;
+    }
+
+    // Draw each lithology layer as a rect with a filter applied
     for (let i = 0; i < lithology.length; i++) {
         const layer = lithology[i];
         const color = layer.colors.length ? layer.colors[0] : LITHOLOGY_COLORS[i % LITHOLOGY_COLORS.length];